#17-Recreate components/correct comments
