+++ client/src/app/component/button/Button.tsx
@@ -2,27 +2,25 @@ import React from "react";
 import clsx from "clsx";
 import styles from "src/app/component/button/Button.module.scss";
 
-interface IButton {
+interface ButtonProps {
   name: string;
-  handleClick: any;
+  handleClick: (event: React.MouseEvent<Element, MouseEvent>) => void;
 }
 
 /**
  * Button component
  */
-const Button: React.FC<IButton> = (props: IButton) => {
+const Button: React.FC<ButtonProps> = (props: ButtonProps) => {
   const btnClassName = clsx(styles.btn);
 
   return (
-    <div>
-      <button
-        className={btnClassName}
-        type="button"
-        onClick={props.handleClick}
-      >
-        {props.name}
-      </button>
-    </div>
+    <button
+      className={btnClassName}
+      type="button"
+      onClick={props.handleClick}
+    >
+      {props.name}
+    </button>
   );
 };
