+++ client/src/app/component/switch/Switch.tsx
@@ -2,15 +2,15 @@ import React from "react";
 import clsx from "clsx";
 import switchStyles from "src/app/component/switch/Switch.module.scss";
 
-interface ISwitch {
+interface SwitchProps {
   isToggle: boolean;
-  onToggle: any;
+  onToggle: (e: React.ChangeEvent<HTMLInputElement>) => void;
 }
 
 /**
  * Switch component
  */
-const Switch: React.FC<ISwitch> = (props: ISwitch) => {
+const Switch: React.FC<SwitchProps> = (props: SwitchProps) => {
   const containerClassName = clsx(switchStyles.container);
   const inputClassName = clsx(switchStyles.container__input);
   const switchClassName = clsx(switchStyles.container__switch);
