+++ client/src/app/logic/errorPage/ErrorPage.tsx
@@ -2,14 +2,10 @@ import React from "react";
 import Paragraph from "src/app/component/paragraph/Paragraph";
 import Headline from "src/app/component/headline/Headline";
 
-const CONTAINER = "container";
-const ERROR_NUMBER = "404";
-const ERROR_TEXT = "Settings Page";
-
-/**
- * Error page
- */
 const ErrorPage: React.FC = () => {
+  const CONTAINER = "container";
+  const ERROR_NUMBER = "404";
+  const ERROR_TEXT = "Settings Page";
   return (
     <section className={CONTAINER}>
       <Headline headline={ERROR_NUMBER} />
