+++ client/src/app/logic/quizPage/QuizPage.tsx
@@ -4,9 +4,6 @@ import Questions from "src/app/component/questions/Questions";
 import clsx from "clsx";
 import styles from "src/app/logic/quizPage/QuizPage.module.scss";
 
-/**
- * Quiz page
- */
 const Quiz: React.FC = () => {
   const questionClassName = clsx(styles.container);
